Leaderboard and end states finally working
