+++ ShowerWithYourDad/Assets/Scenes/End.unity
@@ -131,6 +131,7 @@ GameObject:
   - component: {fileID: 23288547}
   - component: {fileID: 23288546}
   - component: {fileID: 23288548}
+  - component: {fileID: 23288549}
   m_Layer: 5
   m_Name: Panel
   m_TagString: Untagged
@@ -215,12 +216,25 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 76d64f2de80bb9948b5ae052ca3251d1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  leaderboard: {fileID: 0}
+  leaderboard: {fileID: 23288549}
   congratsText: {fileID: 589026035}
   enterInitialsText: {fileID: 2110372257}
   scoreText: {fileID: 468413995}
-  gameStateText: {fileID: 0}
   initialsInput: {fileID: 1023603364}
+--- !u!114 &23288549
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 23288544}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: cd76d262b5bb99a4f936419be3f968df, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  leaderboardEntries: []
+  entriesText: {fileID: 0}
 --- !u!1 &311005375
 GameObject:
   m_ObjectHideFlags: 0
