+++ ShowerWithYourDad/Assets/Scenes/KulaniDuplicate.unity
@@ -810,11 +810,11 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   waves:
-  - waveName: Round 8
+  - waveName: Round 7
     EnemiesInWave:
-    - {fileID: 7096490229995985397, guid: d62a9a30dba2da24aad9017a3a5aef59, type: 3}
-    - {fileID: 3509348608596699017, guid: 976123c0e328cc749a250f6eee07d612, type: 3}
-    - {fileID: 3623852655628012648, guid: c7296e3b6fdfc414cae243365c18a8bc, type: 3}
+    - {fileID: 6731048019753198010, guid: 0baa866252613d54683060a3d0725d3a, type: 3}
+    - {fileID: 8648899589132601297, guid: e1bc6cba4b708a34bb3327a2a660a2e7, type: 3}
+    - {fileID: 7966143991498061257, guid: 3a8348afcd616c946b89655506753e62, type: 3}
     SonsInWave:
     - {fileID: 2820416077298083138, guid: e8530979114d51b4aa87fd0ccc229d16, type: 3}
     - {fileID: 7201847267762199223, guid: d3d8e5088b3f08349be961bf0a724061, type: 3}
@@ -885,19 +885,6 @@ MonoBehaviour:
     TimeBeforeThisWave: 0.2
     roundDuration: 5
   - waveName: Round 6
-    EnemiesInWave:
-    - {fileID: 6731048019753198010, guid: 0baa866252613d54683060a3d0725d3a, type: 3}
-    - {fileID: 8648899589132601297, guid: e1bc6cba4b708a34bb3327a2a660a2e7, type: 3}
-    - {fileID: 7966143991498061257, guid: 3a8348afcd616c946b89655506753e62, type: 3}
-    SonsInWave:
-    - {fileID: 2820416077298083138, guid: e8530979114d51b4aa87fd0ccc229d16, type: 3}
-    - {fileID: 7201847267762199223, guid: d3d8e5088b3f08349be961bf0a724061, type: 3}
-    - {fileID: 3645382543473835879, guid: bb91f36ab22a103479c46d11dca664f5, type: 3}
-    NumberToSpawn: 3
-    NumberSonsToSpawn: 1
-    TimeBeforeThisWave: 0.2
-    roundDuration: 5
-  - waveName: Round 7
     EnemiesInWave:
     - {fileID: 6731048019753198010, guid: 0baa866252613d54683060a3d0725d3a, type: 3}
     - {fileID: 8648899589132601297, guid: e1bc6cba4b708a34bb3327a2a660a2e7, type: 3}
