+++ ShowerWithYourDad/Assets/Scripts/ChocolateCollider1.cs
@@ -1,91 +1,101 @@
 ﻿using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class ChocolateCollider1 : MonoBehaviour
 {
-    public int points = 10;
-    public GameObject ChocolateDad; // Reference to the correct dad
+	public int points = 10;
+	public GameObject ChocolateDad; // Reference to the correct dad
 
-    private bool roundEnded = false;
+	private bool roundEnded = false;
 
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (roundEnded) return;
+	private void OnTriggerEnter2D(Collider2D collision)
+	{
+		if (roundEnded) return;
 
-        Debug.Log($"{gameObject.name} (son) collided with {collision.gameObject.name}");
+		Debug.Log($"{gameObject.name} (son) collided with {collision.gameObject.name}");
 
-        // Handle obstacles first
-        if (collision.CompareTag("Obstacle"))
-        {
-            roundEnded = true;
-            StartCoroutine(Player.instance2.SlowDown());
-            Debug.Log("Hit an obstacle!");
-           // SafeEndRound();
-            return;
-        }
+		// Handle obstacles first
+		if (collision.CompareTag("Obstacle"))
+		{
+			roundEnded = true;
+			//StartCoroutine(Player.instance2.SlowDown());
+			Debug.Log("Hit an obstacle!");
+			// SafeEndRound();
+			return;
+		}
+		//else
+		//{
+			// Check if this is the correct dad
+			if (collision.gameObject == ChocolateDad)
+			{
+				HandleSuccess("Found correct dad!");
+			}
+			// Check if this is a wrong dad (another chocolate)
+			else if (collision.CompareTag("Enemy") || collision.CompareTag("chocolate"))
+			{
+				HandleSuccess("Found correct dad!");
+			}
+			// Anything else
+			else
+			{
+				HandleWrongCollision("Unexpected collision!");
+			}
+		//}
 
-        // Check if this is the correct dad
-        if (collision.gameObject == ChocolateDad)
-        {
-            HandleSuccess("Found correct dad!");
-        }
-        // Check if this is a wrong dad (another chocolate)
-        else if (collision.CompareTag("Enemy") || collision.CompareTag("chocolate"))
-        {
-            HandleSuccess("Found correct dad!");
-        }
-        // Anything else
-        else
-        {
-            HandleWrongCollision("Unexpected collision!");
-        }
-    }
+		//if (roundEnded = true && WaveSpawner.waves.Length == WaveSpawner.currentWaveIndex + 1)
+		//{
+		//	SceneManager.LoadScene("End");
+		//}
+	}
 
-    private void HandleSuccess(string message)
-    {
-        roundEnded = true;
-        if (UIPanel.Instance1 != null)
-        {
-            UIPanel.Instance1.AddScore(points);
-        }
-        if (WaveSpawner.Instance != null)
-        {
-            WaveSpawner.Instance.PlaySound(WaveSpawner.Instance.Coin);
-        }
-        Debug.Log(message);
-        SafeEndRound();
-    }
+	private void HandleSuccess(string message)
+	{
+		roundEnded = true;
+		if (UIPanel.Instance1 != null)
+		{
+			UIPanel.Instance1.AddScore(points);
+		}
 
-    private void HandleWrongCollision(string message)
-    {
-        roundEnded = true;
-        if (WaveSpawner.Instance != null)
-        {
-            WaveSpawner.Instance.PlaySound(WaveSpawner.Instance.WrongHit);
-        }
-        Debug.Log(message);
-        SafeEndRound();
-    }
+		//if (WaveSpawner.Instance != null)
+		{
+			WaveSpawner.Instance.PlaySound(WaveSpawner.Instance.Coin);
+		}
 
-    private void SafeEndRound()
-    {
-        if (Player.instance2 != null)
-        {
-            Player.instance2.EndRoundAndContinue();
-        }
-        else
-        {
-            Debug.LogWarning("Player instance is null, cannot end round");
-            // Use the non-deprecated method
-            Player player = FindAnyObjectByType<Player>();
-            if (player != null)
-            {
-                WaveSpawner.Instance.EmergencyEndRoundAndContinue();
-            }
-        }
-    }
+		Debug.Log(message);
+		SafeEndRound();
+	}
 
-    public void ResetRoundFlag()
-    {
-        roundEnded = false;
-    }
+	private void HandleWrongCollision(string message)
+	{
+		roundEnded = true;
+		if (WaveSpawner.Instance != null)
+		{
+			WaveSpawner.Instance.PlaySound(WaveSpawner.Instance.WrongHit);
+		}
+		Debug.Log(message);
+		SafeEndRound();
+	}
+
+	private void SafeEndRound()
+	{
+		if (Player.instance2 != null)
+		{
+			Player.instance2.EndRoundAndContinue();
+		}
+		else
+		{
+			Debug.LogWarning("Player instance is null, cannot end round");
+			// Use the non-deprecated method
+			Player player = FindAnyObjectByType<Player>();
+			if (player != null)
+			{
+				WaveSpawner.Instance.EmergencyEndRoundAndContinue();
+			}
+		}
+	}
+
+	public void ResetRoundFlag()
+	{
+		roundEnded = false;
+	}
 }
