+++ ShowerWithYourDad/Assets/Scripts/GameManager.cs
@@ -3,23 +3,23 @@ using UnityEngine;
 
 public class GameEvents : MonoBehaviour
 {
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
+	// Start is called once before the first execution of Update after the MonoBehaviour is created
 
-    public static GameEvents instance;
+	public static GameEvents instance;
 
-    private void Awake()
-    {
-        if (instance == null) instance = this;
-        else Destroy(gameObject);
-    }
+	private void Awake()
+	{
+		if (instance == null) instance = this;
+		else Destroy(gameObject);
+	}
 
-    // Events
-    public event Action<int> OnScore;           // Send points
-    public event Action OnRoundSuccess;         // Player found dad
-    public event Action OnRoundFail;            // Wrong or obstacle hit
+	// Events
+	public event Action<int> OnScore;           // Send points
+	public event Action OnRoundSuccess;         // Player found dad
+	public event Action OnRoundFail;            // Wrong or obstacle hit
 
-    // Raise methods
-    public void RaiseScore(int points) => OnScore?.Invoke(points);
-    public void RaiseRoundSuccess() => OnRoundSuccess?.Invoke();
-    public void RaiseRoundFail() => OnRoundFail?.Invoke();
+	// Raise methods
+	public void RaiseScore(int points) => OnScore?.Invoke(points);
+	public void RaiseRoundSuccess() => OnRoundSuccess?.Invoke();
+	public void RaiseRoundFail() => OnRoundFail?.Invoke();
 }
