+++ ShowerWithYourDad/Assets/StreamingAssets/ScoreBoard.txt
@@ -1,6 +1,6 @@
 Rank,Name,Score
-01,---,---
-02,---,---
+01,MAP,0075
+02,MAP,0040
 03,---,---
 04,---,---
 05,---,---
